requirements.txt
# Desenvolvido por (Developed by / 開發者) Felipe Ferreira (f3cdde)

Flask==2.0.2
Flask-SQLAlchemy==2.5.1
Flask-Bcrypt==0.7.1
Flask-Login==0.5.0
python-dotenv==0.19.2
Werkzeug==2.0.3
SQLAlchemy==1.3.23
Flask-WTF==0.14.3
Jinja2==3.0.3
email-validator==1.1.3
Flask-Uploads==0.2.1
Pillow==8.4.0
Flask-Migrate==3.1.0
